refactor to reduce compelxity
